Fixes #58 Added updateOptions parameter and deprecate register-as for ember.js 2
In ember.js 2.4 there is no targetObject available and we cannot set it.
Instead of this we can now pass an attribute to the component that is
observed and when we send to it array of changes it will run those
changes on existing Full Calendar element.

For ember 2.4 and older we can use register-as, but for ember.js 2.0 to
2.4 there will be a message that this method has been deprecated.